rename GPTModel to gpt_model
